Created the framework for powerups
So basically, there is a powerup abstract class. Other powerups extend off this, and implent what happens whent he power up is used. Another important note is that to export all of these powerup ups, I put them inside the index.ts, which means that we can export all the powerups to other files that need them, and if we add new powerups we simply update index.ts, without messing with the other files imports